// OFShopApp/src/model/orderList.ts
export interface OrderList{
    orderId?: any;
    orderStatus: string;
    orderdatetime: string;
    orderLatitude: any;
    orderLongitude: any;
    offercode: string;
    orderLists: any;
    
    wishItemId: string;
    cartId: string;
    userid: string;
    categoryid: string;
    productid: string;  
    cartcount: any;
    item_amount: any;
    cart_amount: any;
    product_name: string;
    product_title: string;
    configuration: string;
    price: any;
    featured_img: string;
    isWishList: boolean;
}